LayoutTests/js/arrowfunction-others-expected.txt

Tests for ES6 arrow function instanceof and typeof operators

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Object.getPrototypeOf(simpleArrowFunction) is Function.prototype
PASS simpleArrowFunction instanceof Function is true
PASS simpleArrowFunction.constructor == Function is true
PASS successfullyParsed is true

TEST COMPLETE
